Use MUnit in ResultMatcherSuite
